+++ Sources/BraintreeCore/URL+IsPayPal.swift
@@ -2,8 +2,9 @@ import Foundation
 
 extension URL {
     
-    /// Used to determine if a the URL is a paypal.com domain in order to format API requests accordingly
+    /// Used to determine if the URL is a paypal.com domain in order to format API requests accordingly
     var isPayPalURL: Bool {
-        absoluteString.contains("api-m.paypal.com")
+        absoluteString.contains(BTCoreConstants.payPalProductionURL.absoluteString) ||
+        absoluteString.contains(BTCoreConstants.payPalSandboxURL.absoluteString)
     }
 }