+++ UnitTests/BraintreeCoreTests/BTAPIClient_Tests.swift
@@ -584,4 +584,26 @@ class BTAPIClient_Tests: XCTestCase {
         let sandboxURL: URL? = apiClient?.graphQLURL(forEnvironment: "unknown")
         XCTAssertEqual(sandboxURL?.absoluteString, "https://payments.braintree-api.com/graphql")
     }
+    
+    func testPayPalBaseURLForEnvironment_returnsSandboxURL() {
+        let apiClientSand = BTAPIClient(authorization: "sandbox_tokenization_key")
+        let baseURLSand: URL? = apiClientSand?.payPalAPIURL(forEnvironment: "sandbox")
+        XCTAssertEqual(baseURLSand?.absoluteString, "https://api.sandbox.paypal.com")
+        
+        let apiClientDev = BTAPIClient(authorization: "sandbox_tokenization_key")
+        let baseURLDev: URL? = apiClientDev?.payPalAPIURL(forEnvironment: "development")
+        XCTAssertEqual(baseURLDev?.absoluteString, "https://api.sandbox.paypal.com")
+    }
+    
+    func testPayPalBaseURLForEnvironment_returnsProductionURL() {
+        let apiClientSand = BTAPIClient(authorization: "production_tokenization_key")
+        let baseURLSand: URL? = apiClientSand?.payPalAPIURL(forEnvironment: "production")
+        XCTAssertEqual(baseURLSand?.absoluteString, "https://api.paypal.com")
+    }
+    
+    func testPayPalBaseURLForEnvironment_returnsProductionURL_asDefault() {
+        let apiClient = BTAPIClient(authorization: "production_tokenization_key")
+        let baseURL: URL? = apiClient?.payPalAPIURL(forEnvironment: "unknown")
+        XCTAssertEqual(baseURL?.absoluteString, "https://api.paypal.com")
+    }
 }