+++ UnitTests/BraintreeCoreTests/BTHTTP_Tests.swift
@@ -488,7 +488,7 @@ final class BTHTTP_Tests: XCTestCase {
     func testPOSTRequests_whenBTHTTPInitializedWithPayPalAPIURL_doesNotSendAuthorizationInBody() {
         let expectation = expectation(description: "POST callback")
 
-        let http = BTHTTP(url: URL(string: "https://api-m.paypal.com")!, authorizationFingerprint: "test-authorization-fingerprint")
+        let http = BTHTTP(url: URL(string: "https://api.paypal.com")!, authorizationFingerprint: "test-authorization-fingerprint")
         http.session = testURLSession
         
         http.post("200.json") { body, response, error in
@@ -498,6 +498,25 @@ final class BTHTTP_Tests: XCTestCase {
 
             let httpRequestBody = BTHTTPTestProtocol.parseRequestBodyFromTestResponseBody(body!)
             XCTAssertFalse(httpRequestBody.contains("authorization_fingerprint"))
+            expectation.fulfill()
+        }
+
+        waitForExpectations(timeout: 2)
+    }
+    
+    func testPOSTRequests_whenBTHTTPInitializedWithPayPalAPIURL_sendsAuthorizationInHeader() {
+        let expectation = expectation(description: "POST callback")
+
+        let http = BTHTTP(url: URL(string: "https://api.paypal.com")!, authorizationFingerprint: "some-fingerprint")
+        http.session = testURLSession
+        
+        http.post("200.json") { body, response, error in
+            XCTAssertNotNil(body)
+            XCTAssertNotNil(response)
+            XCTAssertNil(error)
+
+            let httpRequest = BTHTTPTestProtocol.parseRequestFromTestResponseBody(body!)
+            XCTAssertEqual(httpRequest.allHTTPHeaderFields?["Authorization"], "Bearer some-fingerprint")
             expectation.fulfill()
         }