+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Braintree iOS SDK Release Notes
 
+## unreleased
+* Add BraintreeShopperInsights module (BETA)
+  * Add PrivacyInfo.xcprivacy file
+* BraintreePayPal
+  * Add `BTPayPalRequest.userAuthenticationEmail` optional property
+
 ## 6.14.0 (2024-03-13)
 * BraintreeDataCollector
   * Bump to PPRiskMagnes v5.5.0 with code signing & a privacy manifest file